Move in memory provider to fake_providers dir
