+++ spec/fake_providers/in_memory_provider.rb
